flytt filter-state for tiltakstyper til egen fil
