+++ frontend/mr-admin-flate/src/api/tiltakstyper/useTiltakstyper.ts
@@ -1,7 +1,7 @@
 import { QueryKeys } from "@/api/QueryKeys";
 import { TiltakstyperService } from "@mr/api-client-v2";
 import { useApiSuspenseQuery } from "@mr/frontend-common";
-import { TiltakstypeFilterType } from "../atoms";
+import { TiltakstypeFilterType } from "@/pages/tiltakstyper/filter";
 
 export function useTiltakstyper(filter: TiltakstypeFilterType = {}) {
   const queryFilter = {
